requirements.txt
Django>=4.0
djangorestframework>=3.12
requests>=2.25.1